master Added fallback non-verify decrypt
This change in functionality will now attempt to
verify the signature of a message given the publicKey.

If the public key is null/undefined or the signature
could not be verified for the message, the system will
attempt to simply decrypt the message. If successful,
the unencrypted message is returned, but validSignature
is set to false.

If the message is decrypted and the signature verified,
the unencrypted message is returned and validSignature
is set to true.